batman: Get whole kernel compiling again

+++ sys/compat/linuxkpi/common/include/linux/jiffies.h
@@ -3,6 +3,7 @@
  * Copyright (c) 2010 iX Systems, Inc.
  * Copyright (c) 2010 Panasas, Inc.
  * Copyright (c) 2013-2016 Mellanox Technologies, Ltd.
+ * Copyright (c) 2023 Aymeric Wibo <obiwac@freebsd.org>
  * All rights reserved.
  *
  * Redistribution and use in source and binary forms, with or without
@@ -38,8 +39,8 @@
 #include <sys/kernel.h>
 #include <sys/limits.h>
 
-#define	jiffies			((unsigned long)ticks)
-#define	jiffies_64		((uint64_t)ticks)
+#define	jiffies			ticks
+#define	jiffies_64		ticks
 #define	jiffies_to_msecs(x)     ((unsigned int)(((int64_t)(int)(x)) * 1000 / hz))
 
 #define	MAX_JIFFY_OFFSET	((INT_MAX >> 1) - 1)
